### 2024-04-02 #6571 @Kevin-Mattheus-Moerman
Dear author, @m-d-grunnill, thanks for this submission. I am the AEiC on this track and here to help process the initial steps. Before we proceed, please can you have a look at the following points:

- [ ] The above paper error says:
```
gh issue comment 6571 --body ":warning: [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/8520383968?check_suite_focus=true). Author (Abbas Ghasemi) is missing affiliation."
```
It therefore looks like the affiliation is not formatted properly. Please ensure your paper is formatted properly as per: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography.
- [ ] Please can you elaborate on the key differences/advances of this work with respect to your previous work/publication: https://doi.org/10.1371/JOURNAL.PCBI.1011018 ?
