+++ paper.bib
@@ -3,22 +3,104 @@ Any changes to this file will be lost if it is regenerated by Mendeley.
 
 BibTeX export options can be customized via Options -> BibTeX in Mendeley Desktop
 
+@incollection{Keeling2008c,
+author = {Keeling, Matt J. and Rohani, Pejman},
+booktitle = {Modeling Infectious Diseases in Humans and Animals},
+doi = {10.2307/j.ctvcm4gk0.10},
+isbn = {978-0-691-11617-4},
+pages = {237--240},
+publisher = {Princeton University Press},
+title = {{Metapopulations}},
+year = {2008}
+}
+@article{Almutiry2021,
+abstract = {This paper describes the R package EpiILMCT, which allows users to study the spread of infectious disease using continuous time individual level models (ILMs). The package provides tools for simulation from continuous time ILMs that are based on either spatial demographic, contact network, or a combination of both of them, and for the graphical summarization of epidemics. Model fitting is carried out within a Bayesian Markov Chain Monte Carlo framework. The continuous time ILMs can be implemented within either susceptible-infected-removed (SIR) or susceptible-infected-notified-removed (SIN R) compartmental frameworks. As infectious disease data is often partially observed, data uncertainties in the form of missing infection times – and in some situations missing removal times – are accounted for using data augmentation techniques. The package is illustrated using both simulated and an experimental data set on the spread of the tomato spotted wilt virus disease.},
+archivePrefix = {arXiv},
+arxivId = {2006.00135},
+author = {Almutiry, Waleed and {Vineetha Warriyar}, K. V. and Deardon, Rob},
+doi = {10.18637/jss.v098.i10},
+eprint = {2006.00135},
+file = {:C\:/Users/mdgru/OneDrive - York University/Documents/Mendeley - pdf files/v98i10-1.pdf:pdf},
+issn = {15487660},
+journal = {Journal of Statistical Software},
+keywords = {Contact networks,EpiILMCT,Individual level modeling,Infectious disease,R,Spatial models,contact network,individual level modeling,infectious disease,spatial},
+month = {jun},
+number = {10},
+pages = {1--44},
+publisher = {American Statistical Association},
+title = {{Continuous time individual-level models of infectious disease: Package epiilmct}},
+url = {https://www.jstatsoft.org/index.php/jss/article/view/v098i10},
+volume = {98},
+year = {2021}
+}
+@article{Tye2018,
+abstract = {Ordinary Differential Equations (ODE) are used throughout science where the capture of rates of change in states is sought. While both pieces of commercial and open software exist to study such systems, their efficient and accurate usage frequently requires deep understanding of mathematics and programming. The package we present here, PyGOM, seeks to remove these obstacles for models based on ODE systems. We provide a simple interface for the construction of such systems backed by a comprehensive and easy to use tool--box. This tool--box implements functions to easily perform common operations for ODE systems such as solving, parameter estimation, and stochastic simulation. The package source is freely available and organized in a way that permits easy extension. With both the algebraic and numeric calculations performed automatically (but still accessible), the end user is freed to focus on model development.},
+archivePrefix = {arXiv},
+arxivId = {1803.06934},
+author = {Tye, Edwin and Finnie, Tom and Hall, Ian and Leach, Steve},
+doi = {10.48550/arXiv.1803.06934},
+eprint = {1803.06934},
[…]
