### 2024-04-02 8bfc4cf
Added CONTRIBUTING.md and CODE_OF_CONDUCT.md.
