### 2024-04-02 b019fc2
Improved CONTRIBUTING.md and added reference in README.md.
