### 2024-04-02 #6571 @m-d-grunnill
The first 3 items of the check list have been done.

The main improvements from Grunnill et al (2024) has been to extend the framework to allow for more dimensions in the meta-population structure. The code for that project only allowed for two dimensions of meta-population structure, clusters of people and vaccination statues (as this was groupings of people we were studying).

We have also made some improvements to make the code itself more user friendly:
* The sub-population function for running meta-populations can be given at initialization of MetaCaster. Previously this had to be hard coded on a child class of Base2DMetaPopModel (the old version of MetaCaster).
* Dimensions of a meta-population model can be changed after initialization. This allows for easier comparison of model outputs. Base2DMetaPopModel only allowed dimensions to be set at initialization of its child classes (the old version of MetaCaster.
* Instead of having two separate function for running Latin hypercube sampling in serial and parallel there is now one. This function will perform parallel simulations if given a user defined dask client. The means users can access the dask dashboard to view the progress of simulations run in parallel.
* Events can be made into do nothing events via there own method. Previously this had to be done via a method of the event queue they belonged to.

Other changes made to improve users access to the code.
 * The code can be installed as a python package via pip.
 * Documentation is now available at [https://metacast.readthedocs.io/en/latest/index.html](https://metacast.readthedocs.io/en/latest/index.html)
 * An extensive demo jupyter notebook has been made available [https://github.com/m-d-grunnill/MetaCast/tree/main/demonstrations](https://github.com/m-d-grunnill/MetaCast/tree/main/demonstrations).
