### 2024-04-02 #6571 @m-d-grunnill
> **Submitting author:** @m-d-grunnill ([Martin Grunnill](http://orcid.org/0000-0002-4835-8007)) **Repository:** https://github.com/m-d-grunnill/MetaCast **Branch with paper.md** (empty if default branch): main **Version:** v0.1.5 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Kevin M. Moerman
> ## Status
>
> [![status](https://camo.githubusercontent.com/8814e99c15b5432df6d23c161b4e4c36c6d9568024823552ee9e0c47a1d34fd3/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f66666330663038666433313533666334383762333438323939383130646537642f7374617475732e737667)](https://joss.theoj.org/papers/ffc0f08fd3153fc487b348299810de7d)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @m-d-grunnill. **Currently, there isn't a JOSS editor assigned** to your paper.
>
> @m-d-grunnill if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot @editorialbot is here to help you find and assign reviewers and start the main review. To find out what @editorialbot can do for you type:
>
> ```
> @editorialbot commands
> ```

Could I please suggest Rob Moss as a potential reviewer.
