### 2024-04-04 #6571 @m-d-grunnill
Just checking is it possible the work could be considered under co-publication as outlined in https://joss.readthedocs.io/en/latest/submitting.html . Or have I misunderstood that section. Grunnill et, al. (2024) only describes a specific research use of an earlier version of the code (i.e. modelling COVID-19 at FIFA 2022). The paper I submitted outlines the software and its use more generally.
