### 2024-06-06 #6851 @editorialbot
[REVIEW]: MetaCast: A package for broadCASTing epidemiological and ecological models over META-populations.
**Submitting author:** @m-d-grunnill (<a href="http://orcid.org/0000-0002-4835-8007">Martin Grunnill</a>)
**Repository:** https://github.com/m-d-grunnill/MetaCast
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.6
