### 2024-06-06 #6851 @mstimberg
👋🏼 @m-d-grunnill, @acolum, and @robmoss – this is the review thread for the paper. All of our communications will happen here from now on.

There are additional guidelines in the [message at the start of this issue](https://github.com/openjournals/joss-reviews/issues/6851#issue-2337786134).
