### 2024-06-06 #6851 @acolum
@mstimberg Thank you for setting up this review thread. I have a brief question regarding the JOSS Guidelines. Under the [Substantial Scholarly Effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) section, it is mentioned that: "As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual." However, [the package repository's commit history](https://github.com/m-d-grunnill/MetaCast/commits/main/) appears to be only ~2 months long (from February 6 - April 2, 2024). I was wondering if this suggests that the submission may not fully align with the aforementioned guidelines.
