### 2024-06-06 #6851 @mstimberg
Hi @acolum, thanks for looking into this. If I am not mistaken (please don't hesitate to chime in @m-d-grunnill), the repository continues/builds on work that has been done in other repositories by the same authors (e.g. https://github.com/LIAM-COVID-19-Forecasting/Modelling-Disease-Mitigation-at-Mass-Gatherings-A-Case-Study-of-COVID-19-at-the-2022-FIFA-World-Cup). The comments on top of some source files (e.g. https://github.com/m-d-grunnill/MetaCast/blob/main/src/metacast/infection_seeding.py) also seem to indicate that development started in 2022, so I think the basic criteria for "substantial scholarly effort" are fulfilled. Please also note that the EiC submitted the publication to a "scope check" by the editorial board (mostly due to the potential conflict with the paper associated to the GitHub repository linked above), which [agreed that the submission is in scope for JOSS](https://github.com/openjournals/joss-reviews/issues/6571#issuecomment-2058419462).
