### 2024-06-06 #6851 @m-d-grunnill
As described by @mstimberg the work builds on the code used in the [publication](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1011018). We put the code for that publication [github](https://github.com/LIAM-COVID-19-Forecasting/Modelling-Disease-Mitigation-at-Mass-Gatherings-A-Case-Study-of-COVID-19-at-the-2022-FIFA-World-Cup) for transparency. We made the code much more generalizable, added further quality of life changes and packaged the code up for pip. This newer version is what has been submitted to [JOSS](https://github.com/m-d-grunnill/MetaCast).
