### 2024-06-10 m-d-grunnill/MetaCast#3 @acolum
JOSS Review - Recommended Edits for Paper.md
Hi @m-d-grunnill,

Overall, this paper looks good, but I noticed some grammatical and structural issues:
- Under **Summary**, on line 18, the comma after "models" should be moved to after "flexibly." Also, the acronym "ODE" should be defined before using it.
- Under **Statement of need**:
    - The third word in the title of this section should be capitalized to match the capitalization pattern shown in the other sections' titles.
    - On line 33: The word "to" should be added between "order" and "expedite."
    - On line 37: The phrase that starts with "The dimensions (groupings)..." is a fragment and should be merged with one of the sentences around it.
    - On line 37: The word "cluster" should be "clusters."
    - On lines 38-40: The sentence that starts with "MetaCast builds..." could be rewritten for greater clarity.
    - On lines 42-43: I'd suggest rewriting the phrase "makes MetaCast an ideal package" in the last sentence of this paragraph, as this phrase was already used in the last sentence of the paragraph right above it (under **Summary**).
- Under **State of Field**:
     - On line 46: A comma should be added after "knowledge."
     - On lines 49 and 58: To stay consistent with previous descriptions of the package in the paper, the phrase "for Epidemiology" should be expanded to "for Epidemiology and Ecology."
     - On line 60: `Covasim` should have the same formatting as `Epipack` and the other package names.
     - On line 61: The word "a" should be removed.
     - On line 67: The comma at the end of the line should be a colon.
- Under **Acknowledgements and Funding**, the bullet points are not rendering correctly in the PDF.
