### 2024-06-14 #6851 @m-d-grunnill
This is the first time I have submitted a paper to JOSS. Should I be working on the corrections from the first reviewer, before the 2nd reviewer finishes. OR should I wait until the 2nd reviewer has given their corrections.
