### 2024-06-14 #6851 @mstimberg
@m-d-grunnill there is no clear rule for this on JOSS. Usually, if you agree with the changes suggested by a reviewer, you can apply them right-away. The only situation where I'd strongly recommend waiting for feedback from the other reviewer is if a proposed change is very fundamental. Slightly rewriting the text, fixing grammar/spelling mistakes, etc. definitely do not need a sign-off from the other reviewer. Hope that clears things up, do not hesitate to ping me again if you have further questions.
