### 2024-06-26 m-d-grunnill/MetaCast#3 @robmoss
Hi @m-d-grunnill, I have a few suggestions in addition to those above:

- Under **State of Field**:
  - On line 52: The word "a" should be removed from "Epimode [...] can also perform ~~a~~ sensitivity analyses on these pre-coded models".
  - On lines 53-54: The word "defining" could be removed from the sentence "[...] produce ODE models from ~~defining~~ transitions between", or replaced by "defined" (although I find that reads awkwardly).
  - On line 65: Consider rephrasing "Epimode [...] has agent based modelling", perhaps replacing "has" with "provides" or "implements".
  - On line 69: Similar to my previous suggestion, perhaps say "The Python package `Eir` (Jacob, 2021) **provides** epidemiological models [...]".
