### 2024-06-26 m-d-grunnill/MetaCast#5 @robmoss
Possible typos in example notebook and MetaCast package
Hi @m-d-grunnill, I noticed the following potential typos in the example notebook and the MetaCast package:

- The `MultnomialSeeder` class is missing the "i" in "Multi"?
- The `metacast.event_handling` module provides an `EventQueue` class, but "queue" is spelt "que" in several places, such as:
  - The `metacast.event_handling.event_que` module (unlikely to be an issue, since the `EventQueue` class is typically imported from `metacast.event_handling`)
  - Throughout the demonstration notebook
- "Compliance" is spelt "complaince" in several times in the demonstration notebook.

The last two should be simple to address, while correcting the first one would introduce a breaking change unless the class is exported under both the original and corrected names.

None of these are critical, and they do not detract from the value of the package, the examples presented in the notebook, or the JOSS paper. So please resolve these typos as you see fit, or not at all.
