### 2024-06-26 m-d-grunnill/MetaCast#5 @robmoss
I spotted one more potential typo in §5.2 of the demonstration notebook:

- "In order to generate a LH sample `lhs_prcc` requires a pandas.DataFrame outlining the **bombardiers** of LHS (our parameter certainties)"; I suspect this should be "boundaries"?
