### 2024-06-26 #6851 @robmoss
@m-d-grunnill @mstimberg I've completed my review and filed several issues regarding:

- [Test cases](https://github.com/m-d-grunnill/MetaCast/issues/4);
- [Potential typos](https://github.com/m-d-grunnill/MetaCast/issues/5); and
- [Latin Hypercube sampling](https://github.com/m-d-grunnill/MetaCast/issues/6).

Once these are resolved, I'm happy to recommend accepting this submission :)
