### 2024-07-02 #6851 @m-d-grunnill
Thanks for the review @robmoss and @acolum. I am starting the corrections.

@mstimberg I have started the corrections to the paper.md file. @acolum points out "Under Acknowledgements and Funding, the bullet points are not rendering correctly in the PDF". This renders fine in normal PDF. I am not sure how to fix this?
