### 2024-07-02 d9fcc9f
Fixed grammar in paper.md and typos/mispelling elsewhere.
