+++ paper.md
@@ -50,11 +50,11 @@ bibliography: paper.bib
 
 # Summary
 
-`MetaCast` is a Python package for broadcasting epidemiological and ecological ODE based models
-over metapopulations (structured populations). Users define a function describing the
+`MetaCast` is a Python package for broadcasting epidemiological and ecological Ordinary Differential Equation (ODE)
+based models over metapopulations (structured populations). Users define a function describing the
 subpopulation model. `MetaCast`'s `MetaCaster` then broadcasts the subpopulation model function over dimensions
-of metapopulations. These dimensions can be defined and redefined flexibly allowing for comparisons
-of multidimensional metapopulation models, that can have migration (flows) of populations between
+of metapopulations. These dimensions can be defined and redefined flexibly, allowing for comparisons
+of multidimensional metapopulation models that can have migration (flows) of populations between
 subpopulations. In addition to the metapopulation suite `MetaCast` has several features. A
 multinomial seeder allows users to randomly select infected stages to place an infected
 population in based on the occupancy time of infected states. `MetaCast`'s event queue suite
@@ -63,44 +63,48 @@ parameter values. Sensitivity analysis can be done in `MetaCast` using paralleli
 Partial Rank Correlation Coefficient functions. All of this makes MetaCast an ideal package not only for modelling 
 metapopulations but for wider scenario analysis.
 
-# Statement of need
+# Statement of Need
 
 `MetaCast` was developed from the code base used in a project modelling the spread of 
 COVID-19 at Mass Gathering Events (MGEs), such as the FIFA 2022 World Cup [@Grunnill2024]. 
 During this project there were a number of MGEs that we considered as potential case studies
 before settling on the FIFA 2022 World Cup. As such, even though our epidemiological model 
 was remaining much the same, the resulting change in metapopulation structure between potential case study models
-meant we had to extensively recode the model. In order expedite this recoding due to changes in metapopulation
+meant we had to extensively recode the model. In order to expedite this recoding due to changes in metapopulation
 structure we developed the code in @Grunnill2024. This code allowed us to broadcast our COVID-19 subpopulation model
-over different two-dimensional metapopulations, whilst calculating the force of infections for all subpopulations
-[@Keeling2008c]. The dimensions (groupings) being different cluster of people (subtypes of hosts or visitors) and 
-vaccination status [@Grunnill2024]. `MetaCast` builds upon the code in @Grunnill2024 by allowing for more than two 
-dimensions, that don't require a basis in clusters or vaccination status. `MetaCast` also includes more user-friendly
-versions of the discrete event, sensitivity analyses and infectious population seeding features from @Grunnill2024. 
-This makes MetaCast an ideal package for scenario analyses based around metapopulation models within epidemiology or
-ecology.
+over different two-dimensional metapopulations (based on a clusters of people and their vaccination status [@Grunnill2024]), 
+whilst calculating the force of infections for all subpopulations [@Keeling2008c]. 
+`MetaCast` builds upon and improves the code used in @Grunnill2024. These improvements mean that metapopulations are no 
+longer limited to two dimensions. Furthermore, metapopulation dimensions don't have to be based on clusters of people and their 
+vaccination status. `MetaCast` also includes more user-friendly versions of the discrete event, sensitivity analyses and
+infectious population seeding features from @Grunnill2024. Making MetaCast an ideal package for scenario analyses
+based around metapopulation models within epidemiology or ecology.
 
 # State of Field
 
 There are a number of packages that can be used for epidemiological or ecological modelling accross a number of 
[…]
