### 2024-07-02 #6851 @robmoss
> @mstimberg I have started the corrections to the paper.md file. @acolum points out "Under Acknowledgements and Funding, the bullet points are not rendering correctly in the PDF". This renders fine in normal PDF. I am not sure how to fix this?

@m-d-grunnill I'm not entirely sure how to fix this, but perhaps try inserting a blank line between the preceding sentence ("The authors of this manuscript ...") and the first list item?
