### 2024-07-03 #6851 @mstimberg
> @m-d-grunnill I'm not entirely sure how to fix this, but perhaps try inserting a blank line between the preceding sentence ("The authors of this manuscript ...") and the first list item?

I agree that this is probably an issue of a missing blank line. Please try and let us know if that doesn't fix it.
