### 2024-07-03 a060b9c
Added nox automation for unit tests.
