### 2024-07-03 e794e33
Updated version number and edited paper.md.
