+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "metacast"
-version = "0.1.5"
+version = "0.1.6"
 dependencies = [
     "numpy >= 1.26.3",
     "pandas >= 2.1.4",
