### 2024-07-03 m-d-grunnill/MetaCast#3 @m-d-grunnill
These changes have been made to the paper.md file.
