### 2024-07-03 m-d-grunnill/MetaCast#4 @m-d-grunnill
Thanks for these suggestions. The nox file with github workflow ha been incorporated too.
