### 2024-07-03 m-d-grunnill/MetaCast#5 @m-d-grunnill
These typos have been dealt with.
