### 2024-07-03 m-d-grunnill/MetaCast#6 @m-d-grunnill
This note was in the documentation for the lhs_prcc function as well. But to make this limitation more explicit to users I have put a further note in the README.md file.

I don't think users can currently circumvent this issue using the lhs_prcc function. Therefore, I am opening this up in a new [enhancement issue](https://github.com/m-d-grunnill/MetaCast/issues/8), to keep it separate from the JOSS corrections.
