### 2024-07-03 m-d-grunnill/MetaCast#4 @robmoss
Oh, nice use of a matrix to run the tests against different Python versions!
