### 2024-07-04 d9396a3
Support sampling from arbitrary distributions
This modifies the lhs_prcc function so that it can accept a dictionary
that maps parameter names to percentage-point (inverse-CDF) functions.

See tests/test_lhs_arbitrary_dists.py for an example of drawing samples
from several Beta distributions.
