### 2024-07-05 #6851 @m-d-grunnill
All the issues raised by the reviewers have been addressed and the issues are closed see:
* [JOSS Review - Recommended Edits for Paper.md](https://github.com/m-d-grunnill/MetaCast/issues/3).
* [Cannot run test cases: ModuleNotFoundError](https://github.com/m-d-grunnill/MetaCast/issues/4)
* [Possible typos in example notebook and MetaCast package](https://github.com/m-d-grunnill/MetaCast/issues/5)
* [Latin Hypercube sampling only supports uniform distributions](https://github.com/m-d-grunnill/MetaCast/issues/6)

Being my first submission to JOSS I am unsure of what the next steps are.
