### 2024-07-09 af2992d
Updated versing number in confy.py
