+++ conf.py
@@ -9,7 +9,7 @@
 project = 'MetaCast'
 copyright = '2024, Martin Grunnill'
 author = 'Martin Grunnill'
-release = '0.1.5'
+release = '0.1.6'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
