### 2024-07-10 f8a0aec
Minor punctuation and syntax changes
