+++ paper.md
@@ -55,7 +55,7 @@ based models over metapopulations (structured populations). Users define a funct
 subpopulation model. `MetaCast`'s `MetaCaster` then broadcasts the subpopulation model function over dimensions
 of metapopulations. These dimensions can be defined and redefined flexibly, allowing for comparisons
 of multidimensional metapopulation models that can have migration (flows) of populations between
-subpopulations. In addition to the metapopulation suite `MetaCast` has several features. A
+subpopulations. In addition to the metapopulation suite, `MetaCast` has several features. A
 multinomial seeder allows users to randomly select infected stages to place an infected
 population in, based on the occupancy time of infected states. `MetaCast`'s event queue suite
 can handle discrete events within simulations, such as movement of populations between compartments and changes in 
@@ -71,13 +71,13 @@ During this project there were a number of MGEs that we considered as potential 
 before settling on the FIFA 2022 World Cup. As such, even though our epidemiological model 
 was remaining much the same, the resulting change in metapopulation structure between potential case study models
 meant we had to extensively recode the model. In order to expedite this recoding due to changes in metapopulation
-structure we developed the code in @Grunnill2024. This code allowed us to broadcast our COVID-19 subpopulation model
+structure, we developed the code in @Grunnill2024. This code allowed us to broadcast our COVID-19 subpopulation model
 over different two-dimensional metapopulations (based on clusters of people and their vaccination status [@Grunnill2024]), 
 whilst calculating the force of infections for all subpopulations [@Keeling2008c]. 
 `MetaCast` builds upon and improves the code used in @Grunnill2024. These improvements mean that metapopulations are no 
-longer limited to two dimensions. Furthermore, metapopulation dimensions don't have to be based on clusters of people and their 
+longer limited to two dimensions. Furthermore, metapopulation dimensions do not have to be based on clusters of people and their 
 vaccination status. `MetaCast` also includes more user-friendly versions of the discrete event, sensitivity analyses and
-infectious population seeding features from @Grunnill2024. Making MetaCast an ideal package for scenario analyses
+infectious population seeding features from @Grunnill2024. These features make MetaCast an ideal package for scenario analyses
 based around metapopulation models within epidemiology or ecology.
 
 # State of Field
@@ -100,7 +100,7 @@ modelling. The Python open-source package `Covasim` (COVID-19 Agent-based Simula
 demographic data tailored to specific countries, encompassing age distributions and population sizes, offering 
 sophisticated transmission networks for various social settings (households, schools, workplaces, long-term care 
 facilities, etc). It also incorporates age-specific disease outcomes, viral load dynamics, and a wide array of 
-intervention strategies. R's `EpiMode` [@Jenness2018] implements agent based modelling based around contacts as 
+intervention strategies. R's `EpiMode` [@Jenness2018] implements agent-based modelling based around contacts as 
 discrete events or as a static network model. There are a number of other R epidemiological IBM packages that take 
 spatial or network contact based approaches: `individual` [@Charles2021], `hybridModels` [@Marques2020] and `EpiILMCT` 
 [@Almutiry2021]. The Python package `Eir` [@Jacob2021] provides epidemiological models that incorporate the movements of people. 
