### 2024-07-10 7c71f6f
Fix citations with clashing parentheses
