### 2024-07-10 e9cbbb8
do not spell out URLs in funding acknowledgements
