### 2024-07-10 #6851 @mstimberg
Hi @m-d-grunnill, I made a number of suggestions for (mostly formatting) changes to the paper in m-d-grunnill/MetaCast#10 ­– please merge if you are happy with them.
