### 2024-07-12 #6851 @m-d-grunnill
@mstimberg my co-authors have got back to me.

I have updated the author information to include Jianhong Wu's ORCHID. All other author information is present and correct.
I have made an updated release on GitHub and PyPi at v0.1.6.
I have uploaded a zipped folder of the release at [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.12735031.svg)](https://doi.org/10.5281/zenodo.12735031).
Making sure author details and license are the at [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.12735031.svg)](https://doi.org/10.5281/zenodo.12735031) and https://github.com/m-d-grunnill/MetaCast.
