### 2024-07-13 #6851 @mstimberg
Many thanks for the update @m-d-grunnill , all looks almost good to go from my side as well. Two small things to fix in the Zenodo archive metadata (no need for a new release/archive):

- There is a typo in your last name ("Gunnill" instead of "Grunnil")
- The version is a generic `v1` but should be `v0.1.6` instead.
