### 2024-07-17 #6851 @Kevin-Mattheus-Moerman
- [ ] Please manually edit the ZENODO listed version tag to be `v0.1.6`, i.e. to include the `v` as well.
